refactor(spec_decode): drafter abstraction + ModelRunner aux-unaware
Abstract a Drafter ABC with EagleProposer / DSparkProposer as siblings
(build_drafter factory), and move all aux-hidden-state ownership into the
drafter so ModelRunner is fully aux-unaware.

- Drafter.aux_for(target_hidden_states) reads drafter-owned capture buffers,
  row-aligned to the target hidden states (no token-count juggling at callers).
- EagleProposer.arm_aux_capture unwraps to the aux-owning model to call
  set_aux, then installs a top-level tuple-strip forward hook that copies aux
  into fixed buffers and returns plain hidden. DSpark keeps base layer hooks.
- propose() drops the aux_hidden_states arg; each concrete self-collects.
- ModelRunner drops use_aux_hidden_state_outputs / _aux_hidden_states /
  graph_aux_hidden / _collect_drafter_aux / 3x tuple-unpack / capture-time
  graph_aux; arm_aux_capture runs after the optional TBO wrap.
- UBatchWrapper carries no aux code; capture_tbo_graph just replays its own
  forward hooks inside the captured region.

Renames: eagle.py -> eagle_proposer.py, verify_scheduler.py -> dspark_verify.py,
eagle3_builder.py -> eagle3_kv_builder.py.

Verified: M3 eagle3 GSM8K 0.9371 / accept 71%, V4-Flash DSpark 0.95 / accept 48%,
unit tests 704 passed.
