--- atom/spec_decode/eagle.py
@@ -1,871 +1,0 @@
-import copy
-import logging
-from typing import Optional
-
-import numpy as np
-import torch
-import torch.nn as nn
-from aiter import dtypes
-from aiter.dist.parallel_state import get_pp_group
-from atom.config import CompilationLevel, Config, KVCacheTensor
-from atom.distributed.pcp_utils import (
-    get_pcp_world_size,
-    pcp_allgather_rerange,
-    pcp_pad_dense,
-    pcp_pad_len,
-    pcp_round_robin_split,
-)
-from atom.model_loader.loader import load_model
-from atom.spec_decode.verify_scheduler import VerifyScheduler
-from atom.utils import CpuGpuBuffer, resolve_obj_by_qualname
-from atom.utils import envs
-from atom.utils.forward_context import (
-    DPMetadata,
-    SpecDecodeMetadata,
-    get_forward_context,
-)
-from torch.profiler import record_function
-
-logger = logging.getLogger("atom")
-
-
-support_eagle_model_arch_dict = {
-    "DeepSeekMTPModel": "atom.models.deepseek_mtp.DeepSeekMTP",
-    "DeepseekV4MTPModel": "atom.models.deepseek_v4_mtp.DeepseekV4MTP",
-    "DeepseekV4DSparkModel": "atom.models.deepseek_v4_dspark.DeepseekV4DSpark",
-    "Qwen3NextMTPModel": "atom.models.qwen3_next_mtp.Qwen3NextMTP",
-    "MiMoV2MTPModel": "atom.models.mimo_v2_mtp.MiMoV2MTP",
-    "MiMoV2FlashMTPModel": "atom.models.mimo_v2_mtp.MiMoV2MTP",
-    "Qwen3_5MTPModel": "atom.models.qwen3_5_mtp.Qwen3_5MTP",
-    "Eagle3LlamaModel": "atom.models.eagle3_llama.Eagle3LlamaModel",
-    "Eagle3DeepseekMLAModel": "atom.models.eagle3_deepseek_mla.Eagle3DeepseekMLAModel",
-}
-
-
-def _pcp_active_for_draft_model(draft_model: nn.Module) -> bool:
-    # DeepSeek V2/DSA draft models share this sparse-MLA PCP gate.
-    from atom.models.deepseek_v2 import _pcp_active
-
-    if _pcp_active():
-        return True
-
-    if draft_model.__class__.__name__ != "DeepseekV4MTP":
-        return False
-
-    from atom.models.deepseek_v4 import _pcp_active as _pcp_active_v4
-
-    return _pcp_active_v4()
-
-
-class Eagle3DraftBuilder:
-    """KV cache subsystem for an Eagle3 MHA draft alongside a non-MHA target.
-
-    Implements the same subset of `AttentionMetadataBuilder` hooks that
-    ModelRunner consults during KV pool sizing and per-module binding —
-    `compute_block_bytes`, `allocate_kv_cache_tensors`, and
-    `build_kv_cache_tensor` — so the draft's independent non-MLA cache
-    fits the post-#659 builder protocol without leaking into the target's
-    builder. The draft does NOT drive prepare_decode/prepare_prefill;
-    it piggybacks on the target builder's metadata flow during propose.
-    """
-
-    def __init__(self, model_runner, draft_hf):
-        self.model_runner = model_runner
-        self.draft_hf = draft_hf
-        self.block_size = model_runner.block_size
-        self.num_kv_heads = draft_hf.num_key_value_heads // model_runner.world_size
-        self.num_layers = draft_hf.num_hidden_layers
-        self.head_dim = draft_hf.head_dim
-        self._next_layer_id = 0  # consumed by build_kv_cache_tensor
-        self.num_blocks = 0  # set in allocate_kv_cache_tensors
-
-    def compute_block_bytes(self) -> int:
-        """Per-block bytes for the draft's independent non-MLA KV cache."""
-        kv_dtype_size = dtypes.d_dtypes[
-            self.model_runner.config.kv_cache_dtype
-        ].itemsize
-        bb = (
-            2
-            * self.num_layers
-            * self.block_size
-            * self.num_kv_heads
-            * self.head_dim
-            * kv_dtype_size
-        )
-        if self.model_runner.config.kv_cache_dtype == "fp8":
-            # fp8 KV cache needs an extra per-(layer, block, kv_head) scale
-            # tensor (one fp32 per element) to dequantize fp8 → bf16 at
-            # attention time. Reserve that space alongside the cache.
-            bb += (
-                2
-                * self.num_layers
-                * self.block_size
-                * self.num_kv_heads
-                * dtypes.fp32.itemsize
-            )
-        return bb
-
-    def allocate_kv_cache_tensors(self, num_kv_heads, num_draft_layers) -> dict:
-        """Allocate the draft's [2, L, blocks, block_size, kv_heads, head_dim]
-        cache and matching fp32 scale; ModelRunner setattr's both onto itself
-        under namespaced keys so they don't collide with the target builder's
-        `kv_cache` / `kv_scale`.
-        """
-        runner = self.model_runner
-        config = runner.config
-        # Draft's block budget scales with the target pool: same total token
-        # capacity, just paged at the draft's own block size.
-        self.num_blocks = (
-            config.num_kvcache_blocks * runner.block_size // self.block_size
-        )
-        cache = torch.zeros(
-            2,
-            self.num_layers,
-            self.num_blocks,
-            self.block_size,
-            self.num_kv_heads,
-            self.head_dim,
-            dtype=dtypes.d_dtypes[config.kv_cache_dtype],
-            device="cuda",
-        )
-        scale = torch.zeros(
-            2,
-            self.num_layers,
-            self.num_blocks,
-            self.num_kv_heads,
-            self.block_size,
-            dtype=dtypes.fp32,
-            device="cuda",
-        )
-        logger.info(f"Allocated Eagle3 draft KV cache: {cache.shape}")
-        return {"eagle3_kv_cache": cache, "eagle3_kv_scale": scale}
-
-    def build_kv_cache_tensor(self, layer_id: int, module):
-        """Bind one Eagle3 draft attention module to its slice of the
-        independent draft KV cache. Returns None for non-MHA modules so
-        ModelRunner falls through to the target builder.
-        """
-        if not (
-            hasattr(module, "base_attention")
-            and hasattr(module, "use_mla")
-            and not module.use_mla
-        ):
-            return None
-        runner = self.model_runner
-        idx = self._next_layer_id
-        self._next_layer_id += 1
-        cache = runner.eagle3_kv_cache
-        x = 16 // cache.element_size()
-        k_cache = cache[0, idx].view(
-            self.num_blocks,
-            self.num_kv_heads,
-            self.head_dim // x,
-            self.block_size,
-            x,
-        )
-        v_cache = cache[1, idx].view(
-            self.num_blocks,
-            self.num_kv_heads,
-            self.head_dim,
-            self.block_size,
-        )
-        module.max_model_len = runner.config.max_model_len
-        if runner.config.kv_cache_dtype == "fp8":
-            module.k_scale = runner.eagle3_kv_scale[0, idx]
-            module.v_scale = runner.eagle3_kv_scale[1, idx]
-        module.k_cache = k_cache
-        module.v_cache = v_cache
-        return KVCacheTensor(
-            layer_num=layer_id,
-            k_cache=k_cache,
-            v_cache=v_cache,
-            k_scale=getattr(module, "k_scale", None),
-            v_scale=getattr(module, "v_scale", None),
-        )
-
-    def get_kv_transfer_tensors(self) -> list:
-        from atom.kv_transfer.disaggregation.types import KVTransferRegion
-
-        runner = self.model_runner
-        if not hasattr(runner, "eagle3_kv_cache"):
-            return []
-
-        regions: list[KVTransferRegion] = []
-        cache = runner.eagle3_kv_cache
-        for layer_id in range(self.num_layers):
-            for kv in range(2):
-                t = cache[kv, layer_id]
-                regions.append(
-                    KVTransferRegion(
-                        base_addr=t.data_ptr(),
-                        total_bytes=t.numel() * t.element_size(),
-                        unit_bytes=t.stride(0) * t.element_size(),
-                    )
-                )
-        scale = runner.eagle3_kv_scale
-        if (
-            self.model_runner.config.kv_cache_dtype == "fp8"
-            and scale is not None
-            and scale.numel() > 0
-        ):
-            for layer_id in range(self.num_layers):
-                for kv in range(2):
-                    t = scale[kv, layer_id]
-                    regions.append(
-                        KVTransferRegion(
-                            base_addr=t.data_ptr(),
-                            total_bytes=t.numel() * t.element_size(),
-                            unit_bytes=t.stride(0) * t.element_size(),
-                        )
-                    )
-        return regions
-
-
-class EagleProposer:
-
-    def __init__(
-        self,
-        atom_config: Config,
-        device: torch.device,
-        runner,
-    ):
-        self.config = atom_config
-        self.speculative_config = self.config.speculative_config
-        # DSpark is a parallel block drafter: the verify length is the draft
-        # block size, drafted in a single backbone pass (not mtp_k serial steps).
-        self.use_dspark = bool(
-            getattr(self.speculative_config, "use_dspark", lambda: False)()
-        )
-        if self.use_dspark:
-            draft_cfg = self.speculative_config.draft_model_hf_config
-            self.dspark_block_size = int(getattr(draft_cfg, "dspark_block_size"))
-            # num_speculative_tokens may be unset for DSpark; default to the
-            # full block (Phase 1 uses a static verify length == block size).
-            self.mtp_k: int = (
-                self.speculative_config.num_speculative_tokens or self.dspark_block_size
-            )
-            # Phase 2: confidence-scheduled verification (Level B, variable-length
-            # verify). The ell (per-request verify length) machinery lives in a
-            # reusable VerifyScheduler; propose() feeds it the confidence head
-            # and the next step's calc_spec_decode_metadata consumes the ell map.
-            self.dspark_confidence_schedule = bool(
-                self.config.dspark.confidence_schedule
-            )
-            self.verify_scheduler = (
-                VerifyScheduler(runner) if self.dspark_confidence_schedule else None
-            )
-        else:
-            self.mtp_k: int = self.speculative_config.num_speculative_tokens or 0
-
-        self.runner = runner
-        self.dtype = self.config.torch_dtype
-        self.max_model_len = self.config.max_model_len
-        self.block_size = self.config.kv_cache_block_size
-        self.max_num_tokens = self.config.max_num_batched_tokens
-        self.use_cuda_graph = (
-            self.config.compilation_config.level == CompilationLevel.PIECEWISE
-            and not self.config.enforce_eager
-        )
-        self.cudagraph_batch_sizes = list(
-            reversed(self.config.compilation_config.cudagraph_capture_sizes)
-        )
-
-        self.device = device
-        draft_model_hf_config = self.speculative_config.draft_model_hf_config
-        model_class = resolve_obj_by_qualname(support_eagle_model_arch_dict[draft_model_hf_config.architectures[0]])  # type: ignore
-
-        if self.speculative_config.method == "eagle3":
-            # Eagle3 draft has its own architecture, so build it from the
-            # draft hf_config. Disable torch.compile for the draft to avoid
-            # Dynamo tracing issues with the separate KV cache binding.
-            # Shallow-copy instead of deepcopy: with MLA targets (K2.6), the
-            # atom_config holds non-picklable cuda.Stream objects under
-            # downstream fields that deepcopy can't traverse. We only mutate
-            # hf_config and compilation_config.level on the draft, so
-            # isolating just those two attrs is sufficient.
-            draft_atom_config = copy.copy(atom_config)
-            draft_atom_config.hf_config = draft_model_hf_config
-            draft_atom_config.compilation_config = copy.copy(
-                atom_config.compilation_config
-            )
-            draft_atom_config.compilation_config.level = CompilationLevel.NO_COMPILATION
-            # Draft attention layer_num must continue from the target model's
-            # layer count so it maps to the correct kv_cache_data entry.
-            self.model = model_class(
-                draft_atom_config,
-                layer_offset=atom_config.hf_config.num_hidden_layers,
-            )
-            # MHA draft (e.g. K2.5 LlamaForCausalLMEagle3): owns an independent
-            # non-MLA KV cache via Eagle3DraftBuilder, attached to the runner.
-            # MLA draft (e.g. K2.6 EAGLE 3.1): same MLA shape as target, so
-            # it piggybacks on the target's MLA pool (model_runner accounts
-            # for the +1 draft layer via num_nextn_predict_layers default).
-            draft_is_mla = bool(getattr(draft_model_hf_config, "kv_lora_rank", None))
-            if not draft_is_mla:
-                runner.eagle3_draft_builder = Eagle3DraftBuilder(
-                    runner, draft_model_hf_config
-                )
-        else:
-            self.model = model_class(self.config)
-
-        self._draft_argmax_fused = hasattr(self.model, "compute_draft_token")
-
-        i32_kwargs = {"dtype": torch.int32, "device": self.device}
-        i64_kwargs = {"dtype": torch.int64, "device": self.device}
-        max_bs = self.config.max_num_seqs
-        self.arrange_bs = torch.arange(max_bs + 1, **i32_kwargs)
-        self.cu_num_draft_tokens = CpuGpuBuffer(max_bs, **i32_kwargs)
-        self.target_logits_indices = CpuGpuBuffer(max_bs * self.mtp_k, **i64_kwargs)
-        self.bonus_logits_indices = CpuGpuBuffer(max_bs, **i64_kwargs)
-
-    @staticmethod
-    def _share_if_not_loaded(
-        owner: nn.Module,
-        attr: str,
-        source: nn.Module,
-        loaded: set[str],
-        param_key: str,
-        label: str,
-    ):
-        """Replace *owner.attr* with *source* if the weight was not loaded."""
-        if param_key not in loaded and getattr(owner, attr, None) is not None:
-            logger.info(
-                f"MTP {label} not loaded from checkpoint, "
-                "sharing from the target model."
-            )
-            delattr(owner, attr)
-            setattr(owner, attr, source)
-
-    def load_model(self, target_model: nn.Module) -> None:
-        if self.speculative_config.method == "eagle3":
-            load_model(
-                self.model,
-                self.speculative_config.model,
-                self.speculative_config.draft_model_hf_config,
-                self.config.load_dummy,
-                False,
-            )
-            logger.info(
-                "Eagle3 draft model loaded from %s (independent embed/lm_head)",
-                self.speculative_config.model,
-            )
-            return
-
-        # MTP: load from the target model checkpoint and share embeddings/lm_head.
-        loaded = load_model(
-            self.model,
-            self.config.model,
-            self.speculative_config.draft_model_hf_config,
-            self.config.load_dummy,
-            True,
-        )
-
-        # Resolve the base model (unwrap multimodal wrapper if present)
-        target_base = getattr(target_model, "language_model", target_model)
-
-        # Model-specific share hook escape valve. Models whose embed/lm_head
-        # naming doesn't match the standard `model.embed_tokens` /
-        # `lm_head` convention (e.g. DeepSeek-V4 uses `model.embed` /
-        # `model.head`) implement `share_with_target(target_base)` to do
-        # their own setattr-rebinding and short-circuit the default path.
-        if hasattr(self.model, "share_with_target"):
-            self.model.share_with_target(target_base, loaded)
-            if self.use_dspark and hasattr(self.model, "reset_kv_cache"):
-                # Allocate DSpark's private rolling target-KV window now that the
-                # device/dtype and max concurrency are known.
-                self.model.reset_kv_cache(
-                    self.config.max_num_seqs, self.device, self.dtype
-                )
-            return
-
-        # Share embed_tokens with the target model. Match on the *logical* vocab
-        # (num_embeddings) and hidden dim rather than the stored weight shape, so a
-        # replicated target embed ([vocab, hidden], ATOM_REPLICATE_VOCAB_EMBED) is
-        # still shared onto a TP-sharded draft embed ([vocab/tp, hidden]) — the
-        # draft then reuses the target's replicated table (no post-embed
-        # all-reduce). When both are sharded this is identical to the old check.
-        draft_embed = self.model.model.embed_tokens
-        target_embed = target_base.model.embed_tokens
-        draft_vocab = getattr(draft_embed, "num_embeddings", None)
-        target_vocab = getattr(target_embed, "num_embeddings", None)
-        if (
-            get_pp_group().world_size == 1
-            and draft_vocab is not None
-            and draft_vocab == target_vocab
-            and draft_embed.weight.shape[1] == target_embed.weight.shape[1]
-        ):
-            logger.info(
-                "Assuming the EAGLE head shares the same vocab embedding"
-                " with the target model."
-            )
-            del self.model.model.embed_tokens
-            self.model.model.embed_tokens = target_base.model.embed_tokens
-
-        # Share lm_head from target if not loaded from checkpoint.
-        # Case 1: per-layer shared_head.head (DeepSeek MTP)
-        if hasattr(self.model, "model") and hasattr(self.model.model, "layers"):
-            layers = self.model.model.layers
-            # ModuleDict uses string keys (actual layer indices like "61"),
-            # ModuleList uses integer indices.
-            layer_items = (
-                layers.items() if hasattr(layers, "items") else enumerate(layers)
-            )
-            for key, layer in layer_items:
-                if hasattr(layer, "shared_head"):
-                    self._share_if_not_loaded(
-                        layer.shared_head,
-                        "head",
-                        target_base.lm_head,
-                        loaded,
-                        f"model.layers.{key}.shared_head.head.weight",
-                        "shared_head.head",
-                    )
-        # Case 2: top-level lm_head (Qwen3.5 / Qwen3-Next MTP)
-        self._share_if_not_loaded(
-            self.model,
-            "lm_head",
-            target_base.lm_head,
-            loaded,
-            "lm_head.weight",
-            "lm_head",
-        )
-
-    def _refresh_dp_metadata(self, forward_context, num_local_tokens: int) -> None:
-        parallel_config = self.config.parallel_config
-        if parallel_config.data_parallel_size <= 1:
-            return
-        forward_context.context.dp_uniform_decode = False
-        forward_context.dp_metadata = DPMetadata.make(
-            parallel_config,
-            num_local_tokens,
-        )
-
-    def _propose_dspark(
-        self,
-        *,
-        target_token_ids: torch.Tensor,  # [num_tokens]
-        target_positions: torch.Tensor,  # [num_tokens]
-        num_reject_tokens: torch.Tensor,  # [batch]
-        next_token_ids: torch.Tensor,  # [batch] verified anchor token x0
-        last_token_indices: torch.Tensor,  # [batch] flat index of each anchor row
-        aux_hidden_states: Optional[list[torch.Tensor]],
-    ) -> torch.Tensor:
-        """DSpark block drafting: ONE parallel backbone pass + Markov sampling.
-
-        Unlike the serial Eagle/MTP path (a python loop running the draft model
-        mtp_k times), DSpark generates the whole block in a single forward_spec
-        call. The sequential dependency lives inside the lightweight Markov head,
-        not in repeated heavyweight backbone passes.
-
-        GPU-VERIFY: this path needs an MI3xx run against the reference DSpark to
-        confirm (a) the rolling target-KV window is populated correctly across
-        prefix-cache hits, and (b) the sampled block matches the reference.
-        """
-        forward_context = get_forward_context()
-        context = forward_context.context
-        attn_metadata = forward_context.attn_metadata
-        context.is_draft = True
-        bs = context.batch_size
-
-        if aux_hidden_states is None:
-            raise RuntimeError(
-                "DSpark requires target auxiliary hidden states from "
-                "dspark_target_layer_ids; none were captured."
-            )
-        # Concatenate the configured target layers -> [num_tokens, dim*L].
-        main_hidden_all = torch.cat(aux_hidden_states, dim=-1)
-
-        # Anchor token x0 per request = the just-verified target token, located
-        # at last_token_indices in the flat batch.
-        anchor_ids = next_token_ids
-        anchor_positions = torch.index_select(target_positions, 0, last_token_indices)
-        main_hidden = torch.index_select(main_hidden_all, 0, last_token_indices)
-        state_slot = getattr(attn_metadata, "state_slot_mapping", None)
-        if state_slot is not None:
-            cache_indices = state_slot[:bs].to(torch.long)
-        else:
-            cache_indices = torch.arange(bs, device=anchor_ids.device, dtype=torch.long)
-
-        # Prefill warmup: seed each request's rolling window with the last
-        # min(seq_len, window) target tokens BEFORE drafting. Right after
-        # prefill the window is otherwise empty (only the anchor would be
-        # written), so the first draft block sees almost no target context and
-        # rejects early. Writing the prefill tail lifts first-block acceptance
-        # to the steady-state level. Decode steps skip this (the ring buffer is
-        # already populated from prior steps).
-        if context.is_prefill:
-            cu_seqlens_q = getattr(attn_metadata, "cu_seqlens_q", None)
-            if cu_seqlens_q is not None:
-                window = int(self.model.model.mtp[0].window_size)
-                seqlens = cu_seqlens_q[1 : bs + 1] - cu_seqlens_q[:bs]
-                write_per_batch = int(min(int(seqlens.max().item()), window))
-                self.model.precompute_context_kv(
-                    main_hidden_all,
-                    target_positions,
-                    cache_indices,
-                    cu_seqlens_q=cu_seqlens_q[: bs + 1],
-                    write_per_batch=write_per_batch,
-                )
-
-        # Refresh the rolling target-KV window with the new anchor row, then
-        # draft the block in a single backbone pass.
-        self.model.precompute_context_kv(main_hidden, anchor_positions, cache_indices)
-        # Draft width = the verify horizon mtp_k (num_speculative_tokens). This
-        # may exceed dspark_block_size (the training default); DSpark weights are
-        # draft-width-agnostic so the wider block is drafted in one pass, with
-        # positions past block_size RoPE-extrapolated. Capped at the rolling
-        # window so [window ++ draft] KV stays bounded.
-        window = int(self.model.model.mtp[0].window_size)
-        num_draft = min(self.mtp_k, window)
-        self._refresh_dp_metadata(forward_context, bs * num_draft)
-        draft_token_ids, confidence = self.model.forward_spec(
-            anchor_ids,
-            main_hidden,
-            anchor_positions,
-            cache_indices,
-            num_draft=num_draft,
-        )
-        draft_token_ids = draft_token_ids[:, : self.mtp_k]
-        # Phase 2: confidence-scheduled verification. The hardware-aware prefix
-        # scheduler consumes the confidence head to pick a
-        # per-request verify length ell_r. We compute ell here and stash it; the
-        # actual variable-length verification (Level B) is applied downstream by
-        # truncating each request's scheduled spec tokens to ell_r, which frees
-        # batch capacity instead of the no-op in-block masking of Level A.
-        if self.verify_scheduler is not None and confidence is not None:
-            self.verify_scheduler.set_last_ell(
-                self.verify_scheduler.compute_ell(confidence[:, : self.mtp_k])
-            )
-        elif self.verify_scheduler is not None:
-            self.verify_scheduler.set_last_ell(None)
-        return draft_token_ids
-
-    def propose(
-        self,
-        # [num_tokens]
-        target_token_ids: torch.Tensor,
-        # [num_tokens]
-        target_positions: torch.Tensor,
-        # [num_tokens, hidden_size]
-        target_hidden_states: torch.Tensor,
-        # [batch]
-        num_reject_tokens: torch.Tensor,
-        next_token_ids: torch.Tensor,
-        last_token_indices: torch.Tensor,
-        aux_hidden_states: Optional[list[torch.Tensor]] = None,
-    ) -> torch.Tensor:
-
-        if self.use_dspark:
-            return self._propose_dspark(
-                target_token_ids=target_token_ids,
-                target_positions=target_positions,
-                num_reject_tokens=num_reject_tokens,
-                next_token_ids=next_token_ids,
-                last_token_indices=last_token_indices,
-                aux_hidden_states=aux_hidden_states,
-            )
-
-        forward_context = get_forward_context()
-        context = forward_context.context
-        attn_metadata = forward_context.attn_metadata
-        bs = context.batch_size
-        context.is_draft = True
-
-        assert self.runner is not None
-
-        input_ids = target_token_ids
-        # input_ids[last_token_indices] = next_token_ids
-        input_ids.scatter_(0, last_token_indices, next_token_ids)
-        positions = target_positions + 1
-
-        # Eagle3: project concatenated aux hidden states through fc
-        if aux_hidden_states is not None:
-            concat_aux = torch.cat(aux_hidden_states, dim=-1)
-            hidden_states = self.model.combine_hidden_states(concat_aux)
-        else:
-            hidden_states = target_hidden_states
-
-        draft_token_ids = torch.empty(
-            bs, self.mtp_k, dtype=next_token_ids.dtype, device=next_token_ids.device
-        )
-        if envs.ATOM_DEBUG_FORCE_SKIP_DRAFT_MODEL:
-            draft_token_ids.fill_(-1)
-        var = self.runner.forward_vars
-        target_uses_mla = self.runner.use_mla
-        # Eaale3 only support mha currently
-        draft_uses_mha = hasattr(self.runner, "eagle3_draft_builder")
-
-        # Eagle3 MHA reuses target metadata, but the target may be MLA.  Keep
-        # write slots sized to this draft pass, and when prefix cache is active
-        # restore logical block ids: MLA prefill expands block_tables by
-        # block_ratio for its physical block_size=1 pool, while the draft MHA
-        # cache is indexed by runner.block_size blocks.
-        if draft_uses_mha:
-            attn_metadata.slot_mapping = var["slot_mapping"].gpu[: len(input_ids)]
-            attn_metadata.block_tables = var["block_tables"].gpu[:bs]
-
-        # Backends that expose flat per-seq kv_indices/kv_indptr (MLA, MHA)
-        # wire them through eagle's mid-step block; V4 has block_tables +
-        # context_lens instead (its v4_kv_indices_{swa,csa,hca} are per-token
-        # non-equivalent). Hoisted out of the loop so the value is bound for
-        # every iteration (used at i>=1 too, even though i==0 sets it).
-        has_flat_kv = "kv_indices" in var
-
-        for i in range(self.mtp_k):
-            with record_function(f"draft[{i}/{self.mtp_k} bs={bs}]"):
-                # Re-sync DP token
-                self._refresh_dp_metadata(forward_context, input_ids.shape[0])
-                # ---- Prefill Context Parallel (draft i==0 prefill) --------
-                # The draft's first pass is a prefill that reuses the target's
-                # 1/pcp-reindexed attn_metadata, so it must run on this rank's
-                # 1/pcp query shard (input_ids / positions / previous hidden) and
-                # all-gather the draft hidden back to full token order before the
-                # last-token sampling gather. Later draft steps are decode
-                # (is_prefill False) and run full — identical to the non-PCP path.
-                # `input_ids` / `positions` / `hidden_states` themselves stay full
-                # so the post-i==0 decode-metadata setup (which indexes with the
-                # full `last_token_indices`) is unchanged.
-                pcp_draft_prefill = i == 0 and _pcp_active_for_draft_model(self.model)
-                if pcp_draft_prefill:
-                    pcp_ws = get_pcp_world_size()
-                    n_global_draft = input_ids.shape[0]
-                    n_pad = pcp_pad_len(n_global_draft, pcp_ws) - n_global_draft
-                    d_input_ids = pcp_round_robin_split(
-                        pcp_pad_dense(input_ids, n_pad), pcp_ws
-                    )
-                    d_positions = pcp_round_robin_split(
-                        pcp_pad_dense(positions, n_pad), pcp_ws
-                    )
-                    d_hidden = pcp_round_robin_split(
-                        pcp_pad_dense(hidden_states, n_pad), pcp_ws
-                    )
-                else:
-                    d_input_ids, d_positions, d_hidden = (
-                        input_ids,
-                        positions,
-                        hidden_states,
-                    )
-                ret_hidden_states = self.model(
-                    input_ids=d_input_ids,
-                    positions=d_positions,
-                    hidden_states=d_hidden,
-                )
-                if pcp_draft_prefill:
-                    ret_hidden_states = pcp_allgather_rerange(
-                        ret_hidden_states, pcp_ws
-                    )[:n_global_draft]
-
-                sample_hidden_states = (
-                    torch.index_select(ret_hidden_states, 0, last_token_indices)
-                    if i == 0
-                    else ret_hidden_states
-                )
-                # Distributed argmax (all-gather [N, 2] not [N, vocab]) when the
-                # draft supports it; token-identical to compute_logits().argmax().
-                if self._draft_argmax_fused:
-                    new_draft_ids = self.model.compute_draft_token(sample_hidden_states)
-                else:
-                    logits = self.model.compute_logits(sample_hidden_states)
-                    new_draft_ids = logits.argmax(dim=-1)
-                draft_token_ids[:, i] = new_draft_ids
-
-                if i < self.mtp_k - 1:
-                    do_attn_metadata_update = (
-                        not context.is_prefill
-                        # TODO: FIX this condition after we support3 attention head numbers=32
-                        and self.runner.attn_metadata_builder.num_attention_heads != 32
-                    )
-                    if i == 0:
-                        i0_max_seqlen_q = attn_metadata.max_seqlen_q
-                        attn_metadata.max_seqlen_q = 1
-                        slot_mapping = var["slot_mapping"].gpu[
-                            : bs * attn_metadata.max_seqlen_q
-                        ]
-                        cu_seqlens_q = var["cu_seqlens_q"].gpu[: bs + 1]
-                        attn_metadata.cu_seqlens_q = cu_seqlens_q
-                        attn_metadata.slot_mapping = slot_mapping
-                        if has_flat_kv:
-                            kv_indptr = var["kv_indptr"].gpu[: bs + 1]
-                            kv_indices = var["kv_indices"].gpu
-                            attn_metadata.kv_indptr = kv_indptr
-                            attn_metadata.kv_indices = kv_indices
-                        if target_uses_mla:
-                            kv_last_page_lens = var["kv_last_page_lens"].gpu[:bs]
-                            attn_metadata.kv_last_page_lens = kv_last_page_lens
-                            # Sparse (DSA) MLA decode packs KV per token at
-                            # page_size=1, so it reads the all-1s
-                            # sparse_kv_last_page_lens (NOT the dense per-block
-                            # buffer, which makes the asm kernel over-read past
-                            # the written sparse-index region -> illegal access).
-                            # The draft reuses the target's attn_metadata but
-                            # drops to max_seqlen_q=1, so it must re-point this to
-                            # the per-seq all-1s slice itself.
-                            if "sparse_kv_last_page_lens" in var:
-                                attn_metadata.sparse_kv_last_page_lens = var[
-                                    "sparse_kv_last_page_lens"
-                                ].gpu[:bs]
-                        # block_tables, context_lens, and sparse_kv_indptr are
-                        # needed by both MHA and MLA+sparse attention
-                        attn_metadata.block_tables = var["block_tables"].gpu[:bs]
-                        attn_metadata.context_lens = var["context_lens"].gpu[:bs]
-                        if "sparse_kv_indptr" in var:
-                            attn_metadata.sparse_kv_indptr = var[
-                                "sparse_kv_indptr"
-                            ].gpu[: bs + 1]
-                        cu_seqlens_q[: bs + 1] = self.arrange_bs[: bs + 1]
-                        if target_uses_mla and has_flat_kv:
-                            # MLA: block_size=1, kv_indptr tracks tokens
-                            kv_indptr[1 : bs + 1] -= torch.cumsum(
-                                num_reject_tokens, dim=0
-                            )
-                        if positions.ndim == 1:
-                            positions = torch.index_select(
-                                positions, 0, last_token_indices
-                            )
-                        else:
-                            # MRoPE positions keep the token axis last (e.g.
-                            # [3, num_tokens] for Qwen3.5), so select columns
-                            # instead of indexing dim 0.
-                            positions = torch.index_select(
-                                positions, positions.ndim - 1, last_token_indices
-                            )
-                        context.is_prefill = False
-
-                    # update metadata
-                    attn_metadata.max_seqlen_k += 1
-                    fuse_mtp = positions.ndim == 1 and getattr(
-                        self.runner.attn_metadata_builder,
-                        "fuse_mtp_decode_position_update",
-                        False,
-                    )
-                    if fuse_mtp:
-                        mtp_decode_kwargs = {
-                            "update_context_lens": True,
-                            "positions_out": positions,
-                        }
-                    else:
-                        attn_metadata.context_lens[:bs] += 1
-                        positions += 1
-                        mtp_decode_kwargs = {}
-                    workinfos = self.runner.attn_metadata_builder.prepare_mtp_decode(
-                        bs,
-                        (
-                            attn_metadata.max_seqlen_q
-                            if not do_attn_metadata_update
-                            else i0_max_seqlen_q
-                        ),
-                        attn_metadata.max_seqlen_k,
-                        positions,
-                        only_update=do_attn_metadata_update,
-                        num_reject_tokens=num_reject_tokens if i == 0 else None,
-                        **mtp_decode_kwargs,
-                    )
-                    for k, v in workinfos.items():
-                        attn_metadata.__dict__[k] = v
-                    if has_flat_kv and "slot_mapping" not in workinfos:
-                        # MLA/MHA path: slot derived from flat kv_indices.
-                        slot_mapping[:] = kv_indices[kv_indptr[1 : bs + 1] - 1]
-
-                    input_ids = new_draft_ids
-                    hidden_states = sample_hidden_states
-
-        # self.runner.debug(f"final {draft_token_ids=}")
-        # [batch_size, mtp_k]
-        return draft_token_ids
-
-    def prepare_inputs(
-        self,
-        scheduled_bs: int,
-        # [batch_size]
-        last_token_offset: int | torch.Tensor,
-    ) -> torch.Tensor:
-        forward_context = get_forward_context()
-        attn_metadata = forward_context.attn_metadata
-
-        cu_seqlens_q = attn_metadata.cu_seqlens_q
-        # context_lens = attn_metadata.context_lens
-
-        # Only use decode sequences' context_lens and cu_seqlens_q (num_rejected_tokens length matches decode sequences)
-        # These may contain padding, so we need to slice to match num_rejected_tokens length
-        # context_lens = context_lens[:scheduled_bs]
-        # cu_seqlens_q has length scheduled_bs + 1 (includes 0 at start)
-        cu_seqlens_q = cu_seqlens_q[: scheduled_bs + 1]
-
-        # Calculate new sequence lengths
-        # context_lens += 1
-
-        token_indices = cu_seqlens_q[1:] - last_token_offset
-
-        # Defensive clamp to the valid flat-token range [0, total_tokens-1].
-        # Under DSpark flat-ragged CUDA graph, the drain-phase corner (tiny /
-        # mixed batches) can drive an anchor index just out of range; the anchor
-        # only seeds the DRAFT (a wrong anchor lowers acceptance but never
-        # corrupts the verified/target output — losslessness is preserved), so
-        # clamping is safe and avoids an index_select GPU fault. No-op on the
-        # normal path where indices are already in range.
-
-        if self.use_dspark:
-            upper = (cu_seqlens_q[-1] - 1).clamp_(min=0)
-            token_indices = token_indices.clamp_(min=0)
-            torch.minimum(token_indices, upper, out=token_indices)
-
-        return token_indices
-
-    def calc_spec_decode_metadata(
-        self,
-        num_sampled_tokens: np.ndarray,
-        cu_num_sampled_tokens: np.ndarray,
-        input_ids: torch.Tensor,
-    ) -> SpecDecodeMetadata:
-        scheduled_bs = len(num_sampled_tokens)
-
-        # num_draft = num_sampled - 1 per request. num_sampled_tokens is the
-        # per-seq token count for THIS forward (anchor + drafts). In Phase 1 that
-        # is mtp_k+1 (full). With DSpark plan Y the q-bucket already shrank it to
-        # q (uniform), so deriving num_draft from num_sampled keeps draft / target
-        # / bonus indices consistent with the actual forward layout — no separate
-        # mtp_k constant that could desync (the A-bug: 98%->52%).
-        num_draft_tokens = np.asarray(num_sampled_tokens, dtype=np.int32) - 1
-        np.clip(num_draft_tokens, 0, self.mtp_k, out=num_draft_tokens)
-        sum_drafted_tokens = int(num_draft_tokens.sum())
-
-        # Compute the bonus logits indices.
-        bonus_logits_indices = cu_num_sampled_tokens - 1
-
-        # Compute the draft logits indices.
-        # cu_num_draft_tokens: [3, 3, 5, 5, 6]
-        # arange: [0, 1, 2, 0, 1, 0]
-        cu_num_draft_tokens, arange = self.runner._get_cumsum_and_arange(
-            num_draft_tokens, cumsum_dtype=np.int32
-        )
-        # [0, 0, 0, 5, 5, 9]
-        target_logits_indices = np.repeat(
-            cu_num_sampled_tokens - num_sampled_tokens, num_draft_tokens
-        )
-        # [0, 1, 2, 5, 6, 9]
-        target_logits_indices += arange
-        # self.debug(f"{target_logits_indices=}")
-
-        # Do the CPU -> GPU copy.
-        self.target_logits_indices.np[:sum_drafted_tokens] = target_logits_indices
-        self.cu_num_draft_tokens.np[:scheduled_bs] = cu_num_draft_tokens
-        self.bonus_logits_indices.np[:scheduled_bs] = bonus_logits_indices
-        target_logits_indices = self.target_logits_indices.copy_to_gpu(
-            sum_drafted_tokens
-        )
-        cu_num_draft_tokens = self.cu_num_draft_tokens.copy_to_gpu(scheduled_bs)
-        bonus_logits_indices = self.bonus_logits_indices.copy_to_gpu(scheduled_bs)
-
-        # Compute the draft token ids.
-        # draft_token_indices:      [  1,   2,   3, 105, 106, 208]
-        draft_token_ids = torch.index_select(input_ids[1:], 0, target_logits_indices)
-
-        metadata = SpecDecodeMetadata(
-            draft_token_ids=draft_token_ids,
-            num_spec_steps=self.mtp_k,
-            num_draft_tokens_np=num_draft_tokens,
-            cu_num_draft_tokens=cu_num_draft_tokens,
-            target_logits_indices=target_logits_indices,
-            bonus_logits_indices=bonus_logits_indices,
-        )
-        return metadata
